docs : Swagger 응답 예시 string으로 표시되던 문제 수정
